+++ changelog.md
@@ -1,6 +1,7 @@
 # Next
 - [BUG] Fixed issue with subquery creating `include.where` and a paranoid main model.#2749/#2769
 - UniqueConstraintErrors will now extend from ValidationError making it possible to catch both with `.catch(ValidationError)`
+- [FEATURE] Adds `{save: false}` for belongsTo relationship setters. `user.setOrganization(organization, {save: false})` will then only set the foreign key value, but not trigger a save on `user`.
 
 # 2.0.0-rc4
 - [INTERNALS] Update `inflection` dependency to v1.5.3